Move Spin CLI runtime factors to their own crate
