+++ crates/runtime-config/src/lib.rs
@@ -19,6 +19,7 @@ use spin_factors::runtime_config::toml::GetTomlValue as _;
 use spin_factors::{
     runtime_config::toml::TomlKeyTracker, FactorRuntimeConfigSource, RuntimeConfigSourceFinalizer,
 };
+use spin_runtime_factors::TriggerFactorsRuntimeConfig;
 use spin_sqlite as sqlite;
 use spin_trigger::cli::UserProvidedPath;
 
@@ -153,10 +154,8 @@ where
     }
 }
 
-impl From<ResolvedRuntimeConfig<spin_trigger::TriggerFactorsRuntimeConfig>>
-    for spin_trigger::TriggerFactorsRuntimeConfig
-{
-    fn from(value: ResolvedRuntimeConfig<spin_trigger::TriggerFactorsRuntimeConfig>) -> Self {
+impl From<ResolvedRuntimeConfig<TriggerFactorsRuntimeConfig>> for TriggerFactorsRuntimeConfig {
+    fn from(value: ResolvedRuntimeConfig<TriggerFactorsRuntimeConfig>) -> Self {
         value.runtime_config
     }
 }
@@ -425,7 +424,7 @@ fn sqlite_config_resolver(
     ))
 }
 
-impl TryFrom<TomlRuntimeConfigSource<'_, '_>> for spin_trigger::TriggerFactorsRuntimeConfig {
+impl TryFrom<TomlRuntimeConfigSource<'_, '_>> for TriggerFactorsRuntimeConfig {
     type Error = anyhow::Error;
 
     fn try_from(value: TomlRuntimeConfigSource<'_, '_>) -> Result<Self, Self::Error> {
