+++ crates/trigger/src/cli/stdio.rs
@@ -7,10 +7,10 @@ use std::{
 use anyhow::{Context, Result};
 use spin_common::ui::quoted_path;
 use spin_core::async_trait;
+use spin_factor_wasi::WasiFactor;
+use spin_factors::{HasInstanceBuilder, RuntimeFactors};
 use spin_factors_executor::ExecutorHooks;
 use tokio::io::AsyncWrite;
-
-use crate::factors::TriggerFactors;
 
 /// Which components should have their logs followed on stdout/stderr.
 #[derive(Clone, Debug, Default)]
@@ -89,10 +89,10 @@ impl StdioLoggingExecutorHooks {
 }
 
 #[async_trait]
-impl<U> ExecutorHooks<TriggerFactors, U> for StdioLoggingExecutorHooks {
+impl<F: RuntimeFactors, U> ExecutorHooks<F, U> for StdioLoggingExecutorHooks {
     async fn configure_app(
         &mut self,
-        configured_app: &spin_factors::ConfiguredApp<TriggerFactors>,
+        configured_app: &spin_factors::ConfiguredApp<F>,
     ) -> anyhow::Result<()> {
         self.validate_follows(configured_app.app())?;
         if let Some(dir) = &self.log_dir {
@@ -107,10 +107,12 @@ impl<U> ExecutorHooks<TriggerFactors, U> for StdioLoggingExecutorHooks {
 
     fn prepare_instance(
         &self,
-        builder: &mut spin_factors_executor::FactorsInstanceBuilder<TriggerFactors, U>,
+        builder: &mut spin_factors_executor::FactorsInstanceBuilder<F, U>,
     ) -> anyhow::Result<()> {
         let component_id = builder.app_component().id().to_string();
-        let wasi_builder = builder.factor_builders().wasi();
+        let Some(wasi_builder) = builder.factor_builders().for_factor::<WasiFactor>() else {
+            return Ok(());
+        };
         wasi_builder.stdout_pipe(self.component_stdio_writer(
             &component_id,
             "stdout",
