+++ examples/spin-timer/Cargo.toml
@@ -9,6 +9,7 @@ anyhow = "1.0.68"
 clap = { version = "3.1.15", features = ["derive", "env"] }
 futures = "0.3.25"
 serde = "1.0.188"
+spin-cli = { path = "../.." }
 spin-trigger = { path = "../../crates/trigger" }
 spin-factors = { path = "../../crates/factors" }
 tokio = { version = "1.11", features = ["full"] }
