+++ examples/spin-timer/src/main.rs
@@ -1,9 +1,10 @@
 use anyhow::Error;
 use spin_trigger::cli::FactorsTriggerCommand;
+use spin_cli::runtime_factors::FactorsBuilder;
 
 use trigger_timer::TimerTrigger;
 
-type Command = FactorsTriggerCommand<TimerTrigger, Builder>;
+type Command = FactorsTriggerCommand<TimerTrigger, FactorsBuilder>;
 
 #[tokio::main]
 async fn main() -> Result<(), Error> {
