+++ tests/testing-framework/Cargo.toml
@@ -19,6 +19,7 @@ spin-cli = { path = "../.." }
 spin-factors-executor = { path = "../../crates/factors-executor" }
 spin-http = { path = "../../crates/http" }
 spin-loader = { path = "../../crates/loader" }
+spin-runtime-factors = { path = "../../crates/runtime-factors" }
 spin-trigger = { path = "../../crates/trigger" }
 spin-trigger-http = { path = "../../crates/trigger-http" }
 toml = "0.8.6"
